refactor: 상수 Constants 폴더를 domain 내부로 이동
