+++ src/domain/constants/InputMessage.js
