+++ src/domain/constants/PrinterConstants.js
